# docs/openapi.yaml
openapi: 3.0.4
info:
  title: Online Reservation System API
  version: 2.0.0
  description: |
    API for managing events and seat reservations with asynchronous queue-based processing.
    
    ## Queue-Based Operations
    The `/seats/hold` and `/seats/reserve` endpoints now use asynchronous processing:
    1. Submit request → Immediate response with job ID (202 status)
    2. Poll job status → Check completion status and results
    3. Get final result → Success or failure with detailed information
servers:
  - url: http://localhost:3000
paths:
...
  /seats/job/{jobId}/status:
    get:
      summary: Get the status and result of a queued job
      description: |
        Check the status of a previously queued seat operation. Job results are available 
        for 5 minutes after completion.
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID returned from hold/reserve endpoints
      responses:
        "200":
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/JobStatusProcessing'
                  - $ref: '#/components/schemas/JobStatusCompleted'
                  - $ref: '#/components/schemas/JobStatusFailed'
        "400":
          description: Missing job ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Job not found or still processing
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      jobId:
                        type: string
                        description: The requested job ID
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Seat:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the seat
        eventId:
          type: string
          format: uuid
          description: References the event's ID
        UUID:
          type: string
          description: References the user's UUID (empty if not held)
        status:
          type: string
          enum: ["Available", "On hold", "Reserved"]
          description: Current status of the seat
      required:
        - id
        - eventId
        - UUID
        - status
...
    JobStatusProcessing:
      type: object
      properties:
        status:
          type: string
          enum: ["processing"]
        jobId:
          type: string
          format: uuid
      required:
        - status
        - jobId
      example:
        status: "processing"
        jobId: "d832d651-ba00-48da-88a6-993f39cf8765"
    JobStatusCompleted:
      type: object
      properties:
        status:
          type: string
          enum: ["completed"]
        result:
          type: object
          properties:
            success:
              type: boolean
            message:
              type: string
            seat:
              $ref: '#/components/schemas/Seat'
            status:
              type: integer
              description: HTTP status code
          required:
            - success
            - message
            - status
        completedAt:
          type: integer
          format: int64
          description: Timestamp when job completed
      required:
        - status
        - result
        - completedAt
      example:
        status: "completed"
        result:
          success: true
          message: "Seat held for 60 seconds."
          seat:
            id: "504d8c46-c062-4023-b571-70266e02e0e7"
            eventId: "70dec481-724d-4df8-8d8e-4ee7d3f0b28f"
            UUID: "user-test-123"
            status: "On hold"
          status: 200
        completedAt: 1757366264200
    JobStatusFailed:
      type: object
      properties:
        status:
          type: string
          enum: ["failed"]
        error:
          type: string
          description: Error message explaining why the job failed
        failedAt:
          type: integer
          format: int64
          description: Timestamp when job failed
      required:
        - status
        - error
        - failedAt
      example:
        status: "failed"
        error: "Seat is already locked."
        failedAt: 1757366264200
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error